requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.7.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
Brotli @ file:///croot/brotli-split_1736182456865/work
cachetools==5.3.2
causal-conv1d==1.0.0
certifi==2023.11.17
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
cloudpickle==3.0.0
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
databricks-cli==0.18.0
datasets==2.15.0
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill==0.3.7
docker==6.1.3
docker-pycreds==0.4.0
easydict==1.11
einops==0.7.0
entrypoints==0.4
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.13.1
Flask==3.0.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gmpy2 @ file:///croot/gmpy2_1738085463648/work
google-auth==2.26.2
google-auth-oauthlib==1.2.0
greenlet==3.0.2
grpcio==1.60.0
gunicorn==21.2.0
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==7.0.0
ipython @ file:///croot/ipython_1704833016303/work
itsdangerous==2.1.2
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.1
lmdb==1.4.1
Mako==1.3.0
mamba-ssm==1.0.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
MedPy==0.4.0
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
ml-collections==0.1.1
mlflow==2.9.1
mmcv==1.3.8
mmsegmentation==0.14.1
monai==1.3.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
Pillow==10.1.0
platformdirs==4.1.0
prettytable==3.9.0
progress==1.6
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==4.23.4
psutil==7.0.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments @ file:///croot/pygments_1684279966437/work
PyJWT==2.8.0
pyparsing==3.1.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-hostlist==1.23.0
pytorch-lightning==2.1.3
pytz==2025.2
PyYAML==6.0.1
querystring-parser==1.2.4
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
sentry-sdk==2.27.0
setproctitle==1.3.6
shapely==2.0.2
SimpleITK==2.3.1
six==1.17.0
smmap==5.0.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.4.12
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.1
torchaudio==2.1.1
torchmetrics==1.7.1
torchvision==0.16.1
tqdm==4.66.1
traitlets @ file:///croot/traitlets_1671143879854/work
transformers==4.35.2
triton==2.1.0
typing-inspection==0.4.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
urllib3==2.1.0
wandb==0.19.10
wcwidth==0.2.12
websocket-client==1.7.0
Werkzeug==3.0.1
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
